@@ -1,11 +1,17 @@
 // Code generated by MockGen. DO NOT EDIT.
-// Source: ./machine_factory.go
+// Source: ./pkg/kubevirt/machine_factory.go
+//
+// Generated by this command:
+//
+//	mockgen -source=./pkg/kubevirt/machine_factory.go -destination=./pkg/kubevirt/mock/machine_factory_generated.go -package=mock_kubevirt
+//
 
 // Package mock_kubevirt is a generated GoMock package.
 package mock_kubevirt
 
 import (
 	context "context"
+	net "net"
 	reflect "reflect"
 	time "time"
 
@@ -21,6 +27,7 @@ import (
 type MockMachineInterface struct {
 	ctrl     *gomock.Controller
 	recorder *MockMachineInterfaceMockRecorder
+	isgomock struct{}
 }
 
 // MockMachineInterfaceMockRecorder is the mock recorder for MockMachineInterface.
@@ -63,7 +70,7 @@ func (m *MockMachineInterface) Create(ctx context.Context) error {
 }
 
 // Create indicates an expected call of Create.
-func (mr *MockMachineInterfaceMockRecorder) Create(ctx interface{}) *gomock.Call {
+func (mr *MockMachineInterfaceMockRecorder) Create(ctx any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Create", reflect.TypeOf((*MockMachineInterface)(nil).Create), ctx)
 }
@@ -92,7 +99,7 @@ func (m *MockMachineInterface) DrainNodeIfNeeded(arg0 workloadcluster.WorkloadCl
 }
 
 // DrainNodeIfNeeded indicates an expected call of DrainNodeIfNeeded.
-func (mr *MockMachineInterfaceMockRecorder) DrainNodeIfNeeded(arg0 interface{}) *gomock.Call {
+func (mr *MockMachineInterfaceMockRecorder) DrainNodeIfNeeded(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DrainNodeIfNeeded", reflect.TypeOf((*MockMachineInterface)(nil).DrainNodeIfNeeded), arg0)
 }
@@ -128,7 +135,17 @@ func (mr *MockMachineInterfaceMockRecorder) GenerateProviderID() *gomock.Call {
 
 // GetVMNotReadyReason mocks base method.
 func (m *MockMachineInterface) GetVMNotReadyReason() (string, string) {
-	return "", ""
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetVMNotReadyReason")
+	ret0, _ := ret[0].(string)
+	ret1, _ := ret[1].(string)
+	return ret0, ret1
+}
+
+// GetVMNotReadyReason indicates an expected call of GetVMNotReadyReason.
+func (mr *MockMachineInterfaceMockRecorder) GetVMNotReadyReason() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetVMNotReadyReason", reflect.TypeOf((*MockMachineInterface)(nil).GetVMNotReadyReason))
 }
 
 // IsBootstrapped mocks base method.
@@ -210,6 +227,7 @@ func (mr *MockMachineInterfaceMockRecorder) SupportsCheckingIsBootstrapped() *go
 type MockMachineFactory struct {
 	ctrl     *gomock.Controller
 	recorder *MockMachineFactoryMockRecorder
+	isgomock struct{}
 }
 
 // MockMachineFactoryMockRecorder is the mock recorder for MockMachineFactory.
@@ -230,16 +248,31 @@ func (m *MockMachineFactory) EXPECT() *MockMachineFactoryMockRecorder {
 }
 
 // NewMachine mocks base method.
-func (m *MockMachineFactory) NewMachine(ctx *context0.MachineContext, client client.Client, namespace string, sshKeys *ssh.ClusterNodeSshKeys) (kubevirt.MachineInterface, error) {
+func (m *MockMachineFactory) NewMachine(ctx *context0.MachineContext, arg1 client.Client, namespace string, sshKeys *ssh.ClusterNodeSshKeys) (kubevirt.MachineInterface, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "NewMachine", ctx, client, namespace, sshKeys)
+	ret := m.ctrl.Call(m, "NewMachine", ctx, arg1, namespace, sshKeys)
 	ret0, _ := ret[0].(kubevirt.MachineInterface)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // NewMachine indicates an expected call of NewMachine.
-func (mr *MockMachineFactoryMockRecorder) NewMachine(ctx, client, namespace, sshKeys interface{}) *gomock.Call {
+func (mr *MockMachineFactoryMockRecorder) NewMachine(ctx, arg1, namespace, sshKeys any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NewMachine", reflect.TypeOf((*MockMachineFactory)(nil).NewMachine), ctx, client, namespace, sshKeys)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NewMachine", reflect.TypeOf((*MockMachineFactory)(nil).NewMachine), ctx, arg1, namespace, sshKeys)
+}
+
+// NewMachineWithIPs mocks base method.
+func (m *MockMachineFactory) NewMachineWithIPs(ctx *context0.MachineContext, arg1 client.Client, namespace string, sshKeys *ssh.ClusterNodeSshKeys, allocatedIPs []*net.IPNet) (kubevirt.MachineInterface, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "NewMachineWithIPs", ctx, arg1, namespace, sshKeys, allocatedIPs)
+	ret0, _ := ret[0].(kubevirt.MachineInterface)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// NewMachineWithIPs indicates an expected call of NewMachineWithIPs.
+func (mr *MockMachineFactoryMockRecorder) NewMachineWithIPs(ctx, arg1, namespace, sshKeys, allocatedIPs any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NewMachineWithIPs", reflect.TypeOf((*MockMachineFactory)(nil).NewMachineWithIPs), ctx, arg1, namespace, sshKeys, allocatedIPs)
 }
